+++ CHANGES.md
@@ -3,4 +3,7 @@
 - [NEW] Moved to https://github.com/cloudant/couchbackup repository
 - [NEW] API for using as library is more Node.js like.
 - [CHANGE] The --buffer option is now --buffer-size.
+- [BREAKING CHANGE] The `writeerror` event is now just `error`.
+- [BREAKING CHANGE] The `writecomplete` event is now `finished`.
+- [BREAKING CHANGE] For restoring, the `written` event is now `restored`.
 - [DEPRECATED] Previous API functions.